feat: update inputoptions to v2 and refactor demo
