+++ src/webcomponents.js
@@ -1,4 +1,5 @@
 /* Import any web components used here */
 import '@alaskaairux/ods-button/dist/auro-button';
-import "@alaskaairux/ods-inputoptions";
-import "@alaskaairux/ods-toast";
+import '@alaskaairux/ods-inputoptions/dist/ods-inputoption';
+import '@alaskaairux/ods-inputoptions/dist/ods-inputoption-checkbox-group';
+import '@alaskaairux/ods-toast';
